requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
python-multipart
pandas
icalendar
pymysql
sqlalchemy
holidays
requests
openpyxl
pillow
